Make the 32-bit build correct, not just compilable
Restoring linux/arm/v7 failed the release build. I enabled it on the reasoning
that Go cross-compiles, which was an assumption rather than a measurement.

On 32-bit ARM int is 32 bits and three constants overflow it: noLocation
(4294967295, the geoip "no location" sentinel) and RFC 6996's private-ASN bounds.

noLocation is now typed uint32. It is read from a big-endian uint32 field and
compared against one, so untyped was always the wrong choice and a 64-bit int
merely hid it. In cityindex.go the sentinel check also moves BEFORE the narrowing
to int, because int(raw) wraps negative above 2^31 on a 32-bit build, so checking
after the conversion would have silently stopped matching.

An AS number is 32-bit unsigned, so RemoteAs becomes int64 and parses through a
new safeAS. This is the half that was not merely a compile error: strconv.Atoi
does not truncate a large ASN on a 32-bit build, it fails, so a 4-byte private
ASN would have read as 0 and classified as upstream rather than private.

JSON is unchanged -- int64 marshals identically for these values -- and all 136
gates agree. Verified by cross-compiling amd64, arm64 and arm/v7.

Also: citation-audit no longer scans Changes.md. Pinning the census baseline did
not hold on its own, because the census re-ratchets on every run and the next
local sweep put it straight back. A count that varies with the environment cannot
be ratcheted at all -- the same lesson as the timezone-dependent gates, found the
same day.

+++ internal/collect/routing_test.go
@@ -138,7 +138,7 @@ func TestParseUptime(t *testing.T) {
 func TestClassifyPeer(t *testing.T) {
 	for _, tc := range []struct {
 		name             string
-		as               int
+		as               int64
 		desc, peer, want string
 	}{
 		{"private 16-bit low", 64512, "", "", "private"},
